mcp/requirements.txt
mcp
click>=8.2.0
httpx>=0.27
pydantic>=2.0
pydantic-settings>=2.5.2
requests>=2.32.5
starlette>=0.28.0
uvicorn==0.23.2
fastapi==0.115.12
mcp[cli]==1.8.0
pydantic-ai-slim[mcp]>=0.0.49
